Replace README info with detection of autotools
This will ensure people get informed about the necessary tools and
how to get the package pre-configured without requiring the readme
to be modified when packaged.

In addition, the script will not run autoreconf unilaterally so when
people have an old version it will not attempt to overwrite the
pre-existing configure script.

Should solve #131 and also #93 as well as recent comments on README.md