Reduce notification e-mail spam score.
This adds Content-Disposition and Content-Transfer-Encoding headers to the HTML and text/plain e-mail bodies, thus eliminating one of the issues that tools like Spamassassin check for.

The value `inline` for the Content-Disposition is largely ignored by most e-mail clients these days, but may cause some older clients to display both parts of the message.  It's usage here is consistent with what most e-mail clients set when sending both HTML and text/plain components.

The value `8bit` for the Content-Type-Encoding clarifies that there is unescaped binary data in the text, but that it is limited to 998 character lines and only uses CR and LF at the end of the lines.  `7bit` would work also when dealing with sending the messages in English, and technically be more widely compatible, but we can't assume that the message content will not contain 8-bit characters because certain components (specifically, at least the host name and chart name) may contain Unicode codepoints above 127.

Ideally, we also need to add `charset=` parameters to the Content-Type header for both components, but that requires a bit more work because it's non-trivial to translate the locale's character set to an IANA character set.

Partially fixes #2587