Further reduce spam score by adding encoding info.
A lot of spam filters are more likely to flag email without an encoding
in the Content-Type headers as spam.  Netdata is technically not locale
aware, but the email message is sent from a bash script, and bash is
locale aware, so the message will be encoded in whatever the environment
dictated character encoding is (which is usually UTF-8) these days.

It turns out that my assessment in the previous commit was incorrect,
and it actually is trivial to get a MIME compatible encoding name from a
shell script, so lets try to make sure the encoding is set
automatically.

When not explicitly set by the user, we first try to determine the
system encoding using the `locale` command.  If this fails and the user
has not explicitly set an encoding, then it defaults to UTF-8, which is
the de-facto standard on all modern Linux distributions and thus should
be correct most of the time.

Users can override the detection by setting EMAIL_CHARSET in
health_alarm_notify.conf.

For English speaking users, this should result in no change.  For
non-English speaking users, this may fix display of non-punycode encoded
hostnames.