Add a plugin for Unbound DNS stats.
Unbound is a caching recursive DNS resolver often used to provide a
local DNS cache on Linux systems or to provide DNS service on a router
for a local network.  It provides a wide variety of useful and
interesting statistics through it's remote control interface that can be
collected by tools like Netdata.

This plugin makes a direct connection to the Unbound remote control
interface to pull statistics from it.  Currently, it tracks query rates
(including cache hit and miss rates), info about the internal request
processing list, rudimentary timing info (average and median) for
recursive upstream queries, and optionally info about how many objects
are in each of the various caches.

Most of the settings for the local system can be auto-detected if the
Unbound config file is parseable as YAML (it is technically YAML, but
they allow tabs as indentation, and the default configs shipped by most
distributions use tabs instead of spaces).

This plugin is disabled by default because it requires some non-trivial
changes to the Unbound configuration to configure the control interface.

Supports both TLS encapsulated TCP connections, and local unix sockets.