Add the option to thread email notifications.
This adds a new configuration option in health_alarm_nofity.conf to
insert In-Reply-To and References haders in email notifications.  The
contents of these headers are composed in a deterministic way from the
chart name, alarm name, and host name.

This allows email clients to sort netdata alert emails into groups based
on the combination of chart, alarm, and host, which can help
significantly when dealing with a large number of alerts.  It also
provides something for scripts that might process and filter thes emails
to match on that's a bit easier to parse than the subject line.

We default to not enabling this so that people who have their e-mail
clients configured for threaded reading don't see an unexpected change
in behavior.

Strictly speaking, this is technically a misuse of the headers, as they
are supposed to only reference the Message-ID header of an existing
e-mail.  However, I know of no e-mail clients that enforce standards
compliant behavior (as it would break threading picked up partway
through a thread), so we should be safe.

Minimally tested (using the test-alert.sh script and running for a short
while on a server where I synthetically induced lots of alerts).