Add a module for monitoring logind sessions.
'logind' is a daemon used for tracking active users and sessions on a
system.  By querying this, we can easily get counts of users and
sessions on the system, and additionally differentiate by session type
(graphical, console, or remote).

This module uses the `loginctl` command to query the the service, and
does some dead simple text processing on the output to extract the
required information.  In theory, we could query the DBus methods
directly, but that would get us very little in terms of performance, so
it's not worth the significantly higher complexity.

This module requires zero configuration.  The only thing that is
required is for the user netdata is running as to be able to call
`loginctl list-sessions --no-legend` and get reasonable output back.