Fixed handling of OpenSSL on CentOS/RHEL by bundling a static copy and selecting a configuration directory at install time. (#9263)
* Bundle static OpenSSL in our static builds.

This adds code to bundle a static locally built copy of OpenSSL into our
static builds instead of using the Alpine provided copy. It fixes two
cases where our usage of OpenSSL currently fails:

* On CentOS, RHEL, OEL, Amazon Linux, and their derivatives, the OpenSSL
  configuration directory is `/etc/pki/tls` instead of the normal
  `/etc/ssl`. Any usage of TLS in our static builds currently fails
  there because it can't find that directory.
* TLS usage fails similarly on systems that do not have OpenSSL at all
  for the same reason.

To fix this, the newly bundled copy of OpenSSL is built to use
`/opt/netdata/etc/ssl` as it's configuration directory. This directory
is a symlink created at install time pointing to one of the following
locations (in order of precedence):

* `/etc/pki/tls` (for CentOS, RHEL, OEL, AL, and similar).
* `/etc/ssl` (for sane distros that just use the default path).
* `/opt/netdata/share/ssl` (for systems that don't have OpenSSL, this
  contains a copy of the config and certificates from the build
  environment).

* Ensure other components are built with local OpenSSL.

* Clone directly from the desired tag.