Added support for claiming nodes as part of installation. (#10084)
* Added support for claiming nodes as part of installation.

This adds four new options to the `netdata-installer.sh` script:

* `--claim-token`
* `--claim-rooms`
* `--claim-uri`
* `--claim-proxy`

These directly correspond to the `-token`, `-rooms`, `-uri`, and `-proxy`
options for the `netdata-claim.sh` script. They have the following
associated logic:

* If any are specified and the `--disable-cloud` option is also
  specified, we bail and tell the user to either enable the cloud or
  remove the claiming options.
* If only some but not all of the token, rooms, and uri options  are
  specified, we bail and tell the user that they must pass all three.
* If all three of the token, rooms, and uri are specified, we invoke the
  `netdata-claim.sh` script for the install itself as one of the last
  steps in the installation process, using the values passed to these
  options.

This allows users to directly claim the agent as part of the install,
which is useful for automated installation scenarios.

* Add missing space as suggested by @knatsakis

* Properly handle installs in /.

* Properly handle unprefixed installs.

* Fix another spelling error in an option name.

* Properly fix option naming.

* Move claiming into kickstart script instead of netdata-installer.

This makes us more future-proof.

The required changes also fix some buggy behavior in the option parsing
code in the kickstart scripts.

* Fix checksums.

* Sanely handle the daemon not running during the claiming process.

* Silence incorrect shellcheck warning.

* Simplify condition as suggested by @vkalintiris.

* Clean up old changes that should not be here anymore.

These are leftovers from an earlier revision, they are not actually
needed.

* Add ID generation logic to the claiming script.

This lets it reliably claim nodes which have not yet had the daemon run.

Also fixes a consistency issue in the claiming logic in the Docker
entrypoint.