Reuse the SN_EXISTS bit to track anomaly status. (#11154)
* Replace all usages of SN_EXISTS with SN_DEFAULT_FLAGS.

* Remove references to SN_NOT_EXISTS in comments.

* Replace raw zero constant with SN_EMPTY_SLOT.

* Use get_storage_number_flags only in storage_number.{c,h}

* Compare against SN_EMPTY_SLOT to check if a storage_number exists.

This is safe because:

  1. rrdset_done_interpolate() is the only place where we call store_metric(),
  2. All store_metric() calls, except for one, store an SN_EMPTY_SLOT value.
  3. When we are not storing an SN_EMPTY_SLOT value, the flags that we pass to
     pack_storage_number() can be either SN_EXISTS *or* SN_EXISTS_RESET.

* Compare only the SN_EXISTS_RESET bit to find reset values.

* Remove get_storage_number_flags from storage_number.h

* Do not set storage_number flags outside of rrdset_done_interpolate().

This is a NFC intended to limit the scope of storage_number flags
processing to just one function.

* Set reset bit without overwriting the rest of the flags.

* Rename SN_EXISTS to SN_ANOMALY_BIT.

* Use GOTOs in pack_storage_number to return from a single place.

* Teach pack_storage_number how to handle anomalous zero values.

Up until now, a storage_number had always either the SN_EXISTS or
SN_EXISTS_RESET bit set. This meant that it was not possible for any
packed storage_number to compare equal to the SN_EMPTY_SLOT.

However, the SN_ANOMALY_BIT can be set to zero. This is fine for every
value other than the anomalous 0 value, because it would compare equal to
SN_EMPTY_SLOT. We address this issue by mapping the anomalous zero value
to SN_EXISTS_100 (a number which was not possible to generate with the
previous versions of the agent, ie. it won't exist in older dbengine files).

This change was tested manually by intentionally flipping the anomaly
bit for odd/even iterations in rrdset_done_interpolate. Prior to this
change, charts whose dimensions had 0 values, where showing up in the
dashboard as gaps (SN_EMPTY_SLOT), whereas with this commit the values
are displayed correctly.