Skip building ndsudo when it’s not actually needed. (#18191)
In practice, ndsudo is only actually used by the Go plugin, so we only
need to build and install it if we are building and installing the Go
plugin. This speeds up the build a tiny bit when the Go plugin is not
being built, and removes a non-trivial bit of attack surface from the
eventual install as well.